# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GearUp API
  version: 1.0.0
  description: |
    # 🚀 GearUp REST API

    GearUp is a complete RESTful backend API for an Outdoor Gear Rental Platform.

    ## Features

    - JWT Authentication
    - HTTP-only Cookie Authentication
    - Role-based Authorization
    - Outdoor Gear Rental
    - Stripe Payment Integration
    - Rental Availability Checking
    - Reviews & Ratings
    - Provider Dashboard
    - Admin Dashboard

    ## User Roles

    - Customer
    - Provider
    - Admin

    ## Authentication

    Authentication is handled using **HTTP-only cookies**.

    After a successful login, the server securely stores the access token and refresh token in HTTP-only cookies. Protected endpoints automatically authenticate requests using these cookies.
servers:
  - url: http://localhost:5000
    description: Local Development Server
  - url: https://gearup-1-9p45.onrender.com
    description: Production Server
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Validation failed.
        error:
          type: object
...
    CustomerSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        avatar:
          type: string
          nullable: true
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        customerId:
          type: string
          format: uuid
        gearItemId:
          type: string
          format: uuid
        rentalOrderId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        customer:
          $ref: "#/components/schemas/CustomerSummary"
...
    GearReviewSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "6325e342-7a02-4307-a49d-b2c73cd37aeb"
        name:
          type: string
          example: "Black Diamond Half Dome Helmet Updated"
        image:
          type: string
          nullable: true
          example: "https://images.unsplash.com/photo-1517466787929-bc90951d0974"
    CreateReviewRequest:
      type: object

      required:
        - rentalOrderId
        - gearItemId
        - rating

      properties:
        rentalOrderId:
          type: string
          example: cmrgarh2400001gewj4phq3jz

        gearItemId:
          type: string
          format: uuid
          example: 6325e342-7a02-4307-a49d-b2c73cd37aeb

        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5

        comment:
          type: string
          example: Excellent gear!.
    GearReviewCreateSummary:
      type: object

      properties:
        id:
          type: string
          format: uuid
          example: 6325e342-7a02-4307-a49d-b2c73cd37aeb

        name:
          type: string
          example: Black Diamond Half Dome Helmet Updated
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 35
        totalPage:
          type: integer
          example: 4
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
    UpdateReviewRequest:
      type: object
      minProperties: 1

      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 4

        comment:
          type: string
          example: Excellent gear! The condition was perfect, pickup was smooth, and the provider was very helpful. Highly recommended updated......
    GearReviewGearSummary:
      type: object

      properties:
        id:
          type: string
          format: uuid
          example: 6325e342-7a02-4307-a49d-b2c73cd37aeb

        name:
          type: string
          example: Black Diamond Half Dome Helmet Updated
paths:
...
  /api/reviews:
    post:
      tags:
        - Review

      summary: Create a review

      description: |
        Creates a review for a rented gear item.

        **Authorization Required**

        Only authenticated customers can create reviews.

        The customer is automatically identified from the HTTP-only authentication cookie.

        **Rules**

        - A review can only be created for a rental order that belongs to the authenticated customer.
        - The rental order must have been returned.
        - One review is allowed per customer for each rented gear item.

      operationId: createReview

      security:
        - cookieAuth: []

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'

      responses:
        '201':
          description: Review created successfully.

          content:
            application/json:
              schema:
                type: object

                properties:
                  success:
                    type: boolean
                    example: true

                  statusCode:
                    type: integer
                    example: 201

                  message:
                    type: string
                    example: Review created successfully.

                  data:
                    allOf:
                      - $ref: '#/components/schemas/Review'

                      - type: object

                        properties:
                          customer:
                            $ref: '#/components/schemas/CustomerSummary'

                          gearItem:
                            $ref: '#/components/schemas/GearReviewCreateSummary'
                            

        '400':
          description: Invalid request or review already exists.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '401':
          description: Unauthorized.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '403':
          description: Forbidden. Only customers can create reviews.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '404':
          description: Rental order or gear item not found.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  
    get:
      tags:
        - Review

      summary: Get all reviews

      description: |
        Retrieves all gear reviews with pagination, filtering, sorting,
        customer information, and review statistics.

      operationId: getAllReviews

      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1

        - name: limit
          in: query
          schema:
            type: integer
            default: 10

        - name: rating
          in: query
          description: Filter by rating.
          schema:
            type: integer
            minimum: 1
            maximum: 5

        - name: sortBy
          in: query
          schema:
            type: string
            enum:
              - rating
              - createdAt
            default: createdAt

        - name: sortOrder
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc

      responses:
        '200':
          description: Reviews retrieved successfully.

          content:
            application/json:
              schema:
                type: object

                properties:
                  success:
                    type: boolean
                    example: true

                  statusCode:
                    type: integer
                    example: 200

                  message:
                    type: string
                    example: Reviews retrieved successfully.

                  meta:
                    type: object

                    properties:
                      page:
                        type: integer
                        example: 1

                      limit:
                        type: integer
                        example: 10

                      total:
                        type: integer
                        example: 1

                      totalPage:
                        type: integer
                        example: 1

                      hasNext:
                        type: boolean
                        example: false

                      hasPrev:
                        type: boolean
                        example: false

                  data:
                    type: object

                    properties:
                      statistics:
                        type: object

                        properties:
                          averageRating:
                            type: number
                            format: float
                            example: 4

                          totalReviews:
                            type: integer
                            example: 1

                      reviews:
                        type: array

                        items:
                          allOf:
                            - $ref: '#/components/schemas/Review'

                            - type: object

                              properties:
                                customer:
                                  $ref: '#/components/schemas/CustomerSummary'

                                gearItem:
                                  $ref: '#/components/schemas/GearReviewSummary'

        '500':
          description: Internal server error.
  /api/reviews/me:
    get:
      tags:
        - Review

      summary: Get my reviews

      description: |
        Retrieves the authenticated customer's review history.

        **Authorization Required**

        Only authenticated customers can access this endpoint.

        The customer is automatically identified from the HTTP-only authentication cookie.

      operationId: getMyReviews

      security:
        - cookieAuth: []

      responses:
        '200':
          description: My reviews retrieved successfully.

          content:
            application/json:
              schema:
                type: object

                properties:
                  success:
                    type: boolean
                    example: true

                  statusCode:
                    type: integer
                    example: 200

                  message:
                    type: string
                    example: My reviews retrieved successfully.

                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

                  data:
                    type: array

                    items:
                      type: object

                      allOf:
                        - $ref: '#/components/schemas/Review'

                        - type: object

                          properties:
                            gearItem:
                              $ref: '#/components/schemas/GearReviewSummary'

        '401':
          description: Unauthorized.

        '403':
          description: Only customers can access this resource.
  /api/reviews/gear/{gearItemId}:
    get:
      tags:
        - Review

      summary: Get reviews for a specific gear

      description: |
        Retrieves all reviews for a specific gear item.

        **Authentication**

        This endpoint is public. No authentication is required.

      operationId: getGearReviews

      parameters:
        - name: gearItemId
          in: path
          required: true
          description: Unique gear item ID.
          schema:
            type: string
            format: uuid

        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1

        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1

        - name: rating
          in: query
          description: Filter reviews by rating.
          schema:
            type: integer
            minimum: 1
            maximum: 5

        - name: sortBy
          in: query
          description: Sort reviews by field.
          schema:
            type: string
            enum:
              - rating
              - createdAt

        - name: sortOrder
          in: query
          description: Sort order.
          schema:
            type: string
            enum:
              - asc
              - desc

      responses:
        '200':
          description: Reviews retrieved successfully.

          content:
            application/json:
              schema:
                type: object

                properties:
                  success:
                    type: boolean
                    example: true

                  statusCode:
                    type: integer
                    example: 200

                  message:
                    type: string
                    example: Reviews retrieved successfully.

                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

                  data:
                    type: array

                    items:
                      $ref: '#/components/schemas/Review'

        '404':
          description: Gear item not found.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/reviews/{id}:
    patch:
      tags:
        - Review

      summary: Update a review

      description: |
        Updates an existing review.

        **Authorization Required**

        Only the authenticated customer who created the review can update it.

      operationId: updateReview

      security:
        - cookieAuth: []

      parameters:
        - name: id
          in: path
          required: true
          description: Review ID.
          schema:
            type: string
            format: uuid

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'

      responses:
        '200':
          description: Review updated successfully.

          content:
            application/json:
              schema:
                type: object

                properties:
                  success:
                    type: boolean
                    example: true

                  statusCode:
                    type: integer
                    example: 200

                  message:
                    type: string
                    example: Review updated successfully.

                  data:
                    allOf:
                      - $ref: '#/components/schemas/Review'

                      - type: object
                        properties:
                          customer:
                            $ref: '#/components/schemas/CustomerSummary'

                          gearItem:
                            $ref: '#/components/schemas/GearReviewGearSummary'

        '400':
          description: Invalid request.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '401':
          description: Unauthorized.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '403':
          description: Forbidden.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '404':
          description: Review not found.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Review

      summary: Delete review

      description: |
        Deletes an existing review.

        **Authorization Required**

        Only the authenticated customer who created the review can delete it.

        A customer cannot delete another customer's review.

      operationId: deleteReview

      security:
        - cookieAuth: []

      parameters:
        - name: id
          in: path
          required: true
          description: Review ID.
          schema:
            type: string
            format: uuid

      responses:
        '200':
          description: Review deleted successfully.

          content:
            application/json:
              schema:
                type: object

                properties:
                  success:
                    type: boolean
                    example: true

                  statusCode:
                    type: integer
                    example: 200

                  message:
                    type: string
                    example: Review deleted successfully.

                  data:
                    type: "null"
                    nullable: true
                    example: null

        '401':
          description: Unauthorized.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '403':
          description: Forbidden. Customer does not own this review.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '404':
          description: Review not found.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'